mutt: don't store sent mail in the record
When `copy` is set to true, mutt stores a copy of sent mail in the
record, which is the GMail Sent mailbox. When mbsync is next run,
gmail's copy is synced down to local storage and the local copy from the
record is synced up to GMail, resulting in two copies of every sent
message.

The new configuration relies entirely on gmail to store sent mail.